pyproject.toml
[tool.poetry]
name = "microframe"
version = "1.0.0"
description = "Un micro-framework ASGI inspiré de FastAPI"
authors = ["Traore Eliezer <traoreera@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.13"
starlette = "^0.37.2"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic = {extras = ["email"], version = "^2.12.4"}
pydantic-settings = "^2.2.1"
watchdog = "^4.0.0"
python-multipart = "^0.0.9"
rich = "^13.1.2"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
aioredis = "^2.0.1"
setuptools = "^80.9.0"
jinja2 = "^3.1.6"
annotated-doc = "^0.0.4"
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
black = "^24.3.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
mysql-connector-python = "^9.5.0"
sqlalchemy = "^2.0.44"
autopep8 = "^2.3.2"
isort = "^7.0.0"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
